Removes more of the unnecessary code.

+++ wireguard-manager.sh
@@ -1429,14 +1429,14 @@ PublicKey = ${SERVER_PUBKEY}" >>${WIREGUARD_CLIENT_PATH}/"${NEW_CLIENT_NAME}"-${
       fi
       ;;
     11) # Restore WireGuard Config
+      read -rp "Backup Password: " -e -i "$(cat "${WIREGUARD_BACKUP_PASSWORD_PATH}")" WIREGUARD_BACKUP_PASSWORD
+      if [ -z "${WIREGUARD_BACKUP_PASSWORD}" ]; then
+        exit
+      fi
       if [ -d "${WIREGUARD_PATH}" ]; then
         rm --recursive --force ${WIREGUARD_PATH}
       fi
-      if [ -f "${WIREGUARD_BACKUP_PASSWORD_PATH}" ]; then
-        unzip -P "$(cat "${WIREGUARD_BACKUP_PASSWORD_PATH}")" "${WIREGUARD_CONFIG_BACKUP}" -d "${WIREGUARD_PATH}"
-      else
-        unzip ${WIREGUARD_CONFIG_BACKUP} -d ${WIREGUARD_PATH}
-      fi
+      unzip -P "${WIREGUARD_BACKUP_PASSWORD}" "${WIREGUARD_CONFIG_BACKUP}" -d "${WIREGUARD_PATH}"
       # Restart WireGuard
       if [[ "${CURRENT_INIT_SYSTEM}" == *"systemd"* ]]; then
         systemctl enable --now wg-quick@${WIREGUARD_PUB_NIC}
